tests+observability: green baseline, fail-open visibility, nightly health run
Test suite green for the first time (709 -> 719 pass, 0 fail; was 24 red):
- test_export: fixture now writes the lineage sidecar JSON that export()
  requires for perceiver_heads (8 failures were fixture rot, not bugs)
- test_scripts/test_smoke: V1 script loader falls through to scripts/v1/;
  clip_dedup assertions updated to the current cmd_* API (run_* was renamed)
- test_dataset: _load_vae_latent dtype assertion matched to the documented
  float32 contract (float16 predated the live-VAE-dtype change)
A red baseline trains the reader to ignore red — a 25th failure was invisible.

Fail-open visibility (doctor _check_fail_open_fallbacks): the SREF chain
degrades to legacy silently BY DESIGN, so a run5 whose style step fails every
iteration executes perfectly as run4 with no symptom except ABSENT telemetry.
The doctor now parses the latest done iteration log (canonical
collect_metrics_from_log) and warns on: style_pairing requested but no
style_pair lines; style_pair_pct==0; missing held-out VAL line (champion
ranking silently on the train-batch gap). 72h freshness guard avoids
retro-flagging logs that predate the telemetry.

Live telemetry: trainer heartbeat now carries style_pair_pct per log window
and a final val_done heartbeat with val_cond_gap/val_n_pairs.

Nightly health run (train/scripts/nightly_health.sh + launchd 05:30):
pytest + make mps compile + make test-unit, all GPU-free, verdict JSON at
$DATA_ROOT/health/nightly_health.json surfaced by _check_nightly_health
(stale>48h WARNING, build failure CRITICAL — the QWEN-1 regression shipped
because the mps build was broken three weeks unnoticed). First run: 77s,
all green, validating the bc983b4 C changes via the real build + unit suites.

+15 tests (5 fail-open, 5 nightly-health, plus fixture-repair coverage).

+++ train/tests/test_dataset.py
@@ -292,7 +292,9 @@ class TestLoadVaeLatent:
         self._save_int8(tmpdir, "lat_0", channels=32, lh=64, lw=64)
         result = _load_vae_latent("lat_0", tmpdir)
         assert result is not None
-        assert result.dtype == np.float16
+        # Contract: float32, matching the live VAE encode dtype (the float16
+        # assertion predated that change).
+        assert result.dtype == np.float32
         assert result.shape == (32, 64, 64)
         # 10 * 0.5 = 5.0
         assert np.allclose(result, 5.0, atol=0.01)
